Make Fukuwarai Prototype
