+++ package.json
@@ -26,9 +26,7 @@
     "@monaco-editor/react": "^4.5.1",
     "@radix-ui/react-navigation-menu": "^1.1.2",
     "client-only": "^0.0.1",
-    "css": "^3.0.0",
     "gray-matter": "^4.0.3",
-    "modern-css-reset": "^1.4.0",
     "monaco-editor": "^0.38.0",
     "next": "^13.4.3",
     "next-mdx-remote": "^4.4.1",
@@ -46,14 +44,15 @@
     "server-only": "^0.0.1",
     "sharp": "^0.32.1",
     "shiki": "^0.14.2",
+    "styled-components": "^6.0.0-rc.2",
     "unist-util-visit": "^4.1.2"
   },
   "devDependencies": {
-    "@types/css": "^0.0.33",
     "@types/mapbox__rehype-prism": "^0.8.0",
     "@types/node": "20.1.7",
     "@types/react": "18.2.6",
     "@types/react-dom": "18.2.4",
+    "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.59.6",
     "@typescript-eslint/parser": "^5.59.6",
     "autoprefixer": "^10.4.14",
@@ -75,7 +74,6 @@
     "stylelint-declaration-block-no-ignored-properties": "^2.7.0",
     "stylelint-value-no-unknown-custom-properties": "^4.0.0",
     "tailwindcss": "^3.3.2",
-    "ts-node": "^10.9.1",
     "typescript": "5.0.4"
   },
   "volta": {
